server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [{ name = "Sushant Pandey", email = "sushhantpandey@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "marshmallow (>=4.0.0,<5.0.0)",
    "alembic (>=1.13.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "app" }]

[tool.poetry.scripts]
seed-db = "app.scripts.seed_db:seed_db"
alembic = "alembic.config:main"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 88
target-version = ['py312']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]
combine_as_imports = true
